requirements.txt
SQLAlchemy>=1.3.15
flask-sqlalchemy>=2.4.3
requests>=2.21.0
boto3>=1.13.9
s3fs>=0.4.2
pandas >= 1.0.3
scikit-learn >= 0.23
PyYAML>=5.3.1
pytest>=5.4.3
pymysql>=0.9.3